refactor(git): split git-utils god-module into per-concern modules (#921) (#937)
Decompose the 2327-line src/lib/git/git-utils.ts (37 exports) into
single-concern modules, keeping git-utils.ts as a backward-compatible
re-export barrel so all callers via the barrel are unchanged.

New modules (no circular dependencies):
- git-errors.ts     typed error classes + handleGitApiError +
                    classifyNetworkStderr (credential-safe)
- git-exec.ts       execFile wrappers, write serialization, conflict/
                    network-aware helpers
- git-default-branch.ts  getDefaultBranch / resolveDefaultBranchName
- git-status.ts     porcelain status parsing
- git-log.ts        commit-log / date-range / issue-number extraction
- git-diff.ts       working-tree per-file diff
- git-commit.ts     stage / unstage / commit
- git-branches.ts   list / checkout / create / delete + parse helpers
- git-stash.ts      list / push / pop / apply / drop
- git-reset.ts      reset / revert + default-branch guard
- git-remote.ts     fetch / pull / push

Public API and export signatures are unchanged (behavior-invariant).
The test suite is split to follow the new module boundaries; the
203 original cases are preserved verbatim across 11 per-module files.

Gates: lint / tsc --noEmit / test:unit (7900) / build all green.
